requirements.txt
boto3==1.10.8
psycopg2-binary==2.8.4
